COFFObjectFile.cpp: Appease msvc in r201760.
llvm-svn: 201769

+++ llvm/lib/Object/COFFObjectFile.cpp
@@ -383,7 +383,7 @@ error_code COFFObjectFile::initSymbolTablePtr() {
 
 // Returns the file offset for the given VA.
 error_code COFFObjectFile::getVaPtr(uint32_t Addr, uintptr_t &Res) const {
-  uint32_t ImageBase = PE32Header ? PE32Header->ImageBase : PE32PlusHeader->ImageBase;
+  uint32_t ImageBase = PE32Header ? PE32Header->ImageBase : (uint32_t)PE32PlusHeader->ImageBase;
   return getRvaPtr(Addr - ImageBase, Res);
 }
